Use Gradle version catalog for dependencies
Use the Gradle version catalog feature for dependencies, by
moving libraries, versions and plugins to a TOML file.
Dependencies with multiple artifacts, such as Firebase,
have been bundled. This allows for a more readable dependency
block in the build.gradle file.

Also use the plugins {...} block in favor of the legacy
"apply plugin". This allows us to skip the classpath
declaration in the root build.gradle file.

Fixes #141.